Fixed error codes data types

+++ Assets/Plugins/StreamChat/Core/DTO/Models/APIErrorDTO.cs
@@ -20,13 +20,13 @@ namespace StreamChat.Core.DTO.Models
         /// Response HTTP status code
         /// </summary>
         [Newtonsoft.Json.JsonProperty("StatusCode", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
-        public double? StatusCode { get; set; }
+        public int? StatusCode { get; set; }
 
         /// <summary>
         /// API error code
         /// </summary>
         [Newtonsoft.Json.JsonProperty("code", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
-        public double? Code { get; set; }
+        public int? Code { get; set; }
 
         /// <summary>
         /// Request duration
